requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# manage.py: 8
# saleor/cart/__init__.py: 3
# saleor/cart/forms.py: 5
# saleor/cart/models.py: 7,8,9,10,11
# saleor/cart/urls.py: 1
# saleor/cart/utils.py: 6,7,8,9
# saleor/cart/views.py: 3,4,5,6
# saleor/celeryconf.py: 3
# saleor/checkout/core.py: 5,6,7,8,9
# saleor/checkout/forms.py: 2,3,4
# saleor/checkout/urls.py: 1
# saleor/checkout/views/__init__.py: 1,2
# saleor/checkout/views/discount.py: 4,5,6,7,8
# saleor/checkout/views/shipping.py: 1,2
# saleor/checkout/views/summary.py: 4
# saleor/checkout/views/validators.py: 3,4
# saleor/core/__init__.py: 2
# saleor/core/analytics.py: 4
# saleor/core/context_processors.py: 4,5,6
# saleor/core/forms.py: 3
# saleor/core/management/commands/populatedb.py: 1,2,3,4
# saleor/core/permissions.py: 1
# saleor/core/sitemaps.py: 1
# saleor/core/storages.py: 1
# saleor/core/templatetags/attributes.py: 1
# saleor/core/templatetags/jsonld.py: 3
# saleor/core/templatetags/markdown.py: 1,2
# saleor/core/templatetags/materializecss.py: 1,2,3
# saleor/core/templatetags/shop.py: 3,4,5
# saleor/core/templatetags/status.py: 1
# saleor/core/templatetags/version.py: 1
# saleor/core/urls.py: 1
# saleor/core/utils/__init__.py: 8
# saleor/core/views.py: 1,2,3,4,5
# saleor/dashboard/category/filters.py: 1
# saleor/dashboard/category/forms.py: 1,2,3,4
# saleor/dashboard/category/urls.py: 1
# saleor/dashboard/category/views.py: 4,5,6,7
# saleor/dashboard/customer/filters.py: 1,2,3
# saleor/dashboard/customer/urls.py: 1
# saleor/dashboard/customer/views.py: 5,6,7
# saleor/dashboard/order/filters.py: 1,2,3
# saleor/dashboard/order/forms.py: 5
# saleor/dashboard/order/urls.py: 1
# saleor/dashboard/order/utils.py: 1,2,3
# saleor/dashboard/order/views.py: 5,6,7,8
# saleor/dashboard/product/__init__.py: 1
# saleor/dashboard/product/filters.py: 1,2
# saleor/dashboard/product/forms.py: 1,2,3,4,5,6,7
# saleor/dashboard/product/urls.py: 1
# saleor/dashboard/product/views.py: 7,8
# saleor/dashboard/product/widgets.py: 1
# saleor/dashboard/search/urls.py: 1
# saleor/dashboard/search/views.py: 1,2,3
# saleor/dashboard/staff/emails.py: 2,3,4,5,6
# saleor/dashboard/staff/filters.py: 1,2,3
# saleor/dashboard/staff/forms.py: 1
# saleor/dashboard/staff/urls.py: 1
# saleor/dashboard/staff/views.py: 3,4,5
# saleor/dashboard/templatetags/chips.py: 3,4
# saleor/dashboard/templatetags/utils.py: 3,4
# saleor/dashboard/urls.py: 1
# saleor/dashboard/views.py: 6,7
# saleor/dashboard/widgets.py: 1,2
# saleor/graphql/utils.py: 1
# saleor/order/__init__.py: 1
# saleor/order/emails.py: 2,3
# saleor/order/models.py: 6,7,8,9,10
# saleor/order/urls.py: 1
# saleor/order/utils.py: 5,6
# saleor/order/views.py: 3,4
# saleor/product/__init__.py: 1
# saleor/product/filters.py: 3,4
# saleor/product/forms.py: 3,4,5
# saleor/product/models.py: 8,9,10,11,12
# saleor/product/templatetags/product_images.py: 5,6,7
# saleor/product/urls.py: 1
# saleor/product/utils.py: 4,5,6
# saleor/product/views.py: 6,7,8,9
# saleor/registration/forms.py: 1,2,3
# saleor/registration/urls.py: 1,2
# saleor/registration/views.py: 2,3,4,5
# saleor/search/backends/picker.py: 1,2
# saleor/search/backends/postgresql_dashboard.py: 1
# saleor/search/backends/postgresql_storefront.py: 2,3
# saleor/search/forms.py: 1,2
# saleor/search/urls.py: 1
# saleor/search/views.py: 2,3,4
# saleor/settings.py: 6,250
# saleor/site/context_processors.py: 1
# saleor/site/models.py: 1,2,3
# saleor/site/patch_sites.py: 8,9,10,18
# saleor/site/utils.py: 1
# saleor/urls.py: 2,3,4,5,6
# saleor/userprofile/management/commands/changepassword.py: 1
# saleor/userprofile/management/commands/createsuperuser.py: 1
# saleor/userprofile/models.py: 1,3,4,5,6
# saleor/wsgi/__init__.py: 22
Django == 1.11.5

# saleor/core/templatetags/markdown.py: 3
Markdown == 2.6.9

# saleor/dashboard/order/utils.py: 22
WeasyPrint == 0.41

# saleor/celeryconf.py: 2
# saleor/dashboard/staff/emails.py: 1
# saleor/order/emails.py: 1
celery == 4.1.0

# saleor/settings.py: 4
dj_database_url == 0.4.2

# saleor/settings.py: 5
dj_email_url == 0.0.10

# saleor/core/utils/form_renderer.py: 1
django_bootstrap3 == 9.0.0

# saleor/settings.py: 7
django_cache_url == 1.3.1

# saleor/search/documents.py: 1
django_elasticsearch_dsl == 0.4.1

# saleor/core/filters.py: 1
# saleor/core/templatetags/materializecss.py: 4
# saleor/dashboard/category/filters.py: 2
# saleor/dashboard/customer/filters.py: 4
# saleor/dashboard/order/filters.py: 4
# saleor/dashboard/product/filters.py: 3
# saleor/dashboard/staff/filters.py: 4
# saleor/dashboard/templatetags/utils.py: 5
# saleor/dashboard/widgets.py: 3
# saleor/product/filters.py: 5
django_filter == 1.0.4

# saleor/product/models.py: 13,14
django_mptt == 0.8.7

# saleor/core/storages.py: 2
django_storages == 1.6.5

# saleor/dashboard/staff/emails.py: 7
# saleor/order/emails.py: 4
django_templated_email == 2.2.0

# saleor/dashboard/product/widgets.py: 2
# saleor/dashboard/templatetags/utils.py: 6
# saleor/product/models.py: 17
django_versatileimagefield == 1.7.1

# saleor/search/backends/elasticsearch_dashboard.py: 1
# saleor/search/backends/elasticsearch_storefront.py: 1
# saleor/search/documents.py: 2
elasticsearch_dsl == 5.3.0

# saleor/core/analytics.py: 3
google_measurement_protocol == 0.1.6

# saleor/graphql/scalars.py: 1
# saleor/graphql/utils.py: 2
graphene == 2.0.1

# saleor/urls.py: 7
graphene_django == 2.0.0

# saleor/graphql/scalars.py: 2
graphql_core == 2.0

# saleor/cart/models.py: 12
jsonfield == 2.0.2

# saleor/cart/forms.py: 6
# saleor/cart/models.py: 13
# saleor/checkout/views/summary.py: 5
# saleor/order/models.py: 11
# saleor/product/models.py: 15
satchless == 1.1.3

# node_modules/node-gyp/gyp/setup.py: 7
setuptools == 36.5.0

# saleor/dashboard/category/forms.py: 5
# saleor/product/models.py: 16
text_unidecode == 1.0
